Try find suitable range for further testing

+++ src/main/java/examples/Example_2.java
@@ -1,7 +1,7 @@
 package examples;
 
 public class Example_2 {
-    public static void start(int n) {
+    public static void start(long n) {
         long c = 0;
         for (int i = 0; i < n; i++) {
             for (int j = 0; j < n; j++) {
